fix: directory handle leak and uint32 overflow
- torrent-files.cc: wrap dir handle in RAII guard to prevent leak
  if recursive depth_first_walk() throws an exception. Previously
  the tr_sys_dir_close() call would be skipped during stack unwinding.
- completion.cc: cast begin_block to uint64_t before adding 1 to
  prevent uint32_t overflow for torrents approaching 64 TiB with
  16 KiB blocks (max ~2^32 blocks).

Notes: Fixed potential resource leak on exception paths during
torrent file enumeration. Fixed integer overflow in completion
size calculation for very large torrents.

+++ libtransmission/completion.cc
@@ -213,7 +213,7 @@ uint64_t tr_completion::count_has_bytes_in_span(tr_byte_span_t span) const
     // the first block
     if (has_block(begin_block))
     {
-        uint64_t u = begin_block + 1;
+        uint64_t u = static_cast<uint64_t>(begin_block) + 1;
         u *= tr_block_info::BlockSize;
         u -= begin_byte;
         total += u;
